requirements.txt
openai-whisper==20231117
moviepy==1.0.3
ffmpeg-python==0.2.0
SQLAlchemy==2.0.27
configparser==5.3.0
python-dotenv==1.0.1
tqdm==4.66.2
colorlog==6.7.0
openai==1.10.0
pandas==2.0.3
openpyxl==3.1.2